+++ hal/src/sercom/uart.rs
@@ -381,6 +381,30 @@ let (chan1, rx, rx_buffer) = rx_dma.wait();
 
 "
 )]
+//! # `async` operation
+//!
+//! When the `async` Cargo feature is enabled, a [`Uart`] can be used for
+//! `async` operations. Configuring a [`Uart`] in async mode is relatively
+//! simple:
+//!
+//! * Bind the corresponding `SERCOM` interrupt source to the UART
+//!   [`InterruptHandler`] (refer to the module-level [`async_hal`]
+//!   documentation for more information).
+//! * Turn a previously configured [`Uart`] into a [`UartFuture`] by calling
+//!   [`Uart::into_future`]
+//! * Optionally, add DMA channels to RX, TX or both using
+//!   [`UartFuture::with_rx_dma_channel`] and
+//!   [`UartFuture::with_tx_dma_channel`]. The API is exactly the same whether
+//!   DMA channels are used or not.
+//! * Use the provided async methods for reading or writing to the UART
+//!   peripheral.
+//!
+//! `UartFuture` implements `AsRef<Uart>` and `AsMut<Uart>` so
+//! that it can be reconfigured using the regular [`Uart`] methods. It also
+//! exposes a [`split`](UartFuture::split) method to split it into its RX and TX
+//! parts.
+//!
+//! [`async_hal`]: crate::async_hal
 
 use atsamd_hal_macros::{hal_cfg, hal_module};
 
@@ -405,6 +429,11 @@ mod config;
 pub use config::*;
 
 pub mod impl_ehal;
+
+#[cfg(feature = "async")]
+mod async_api;
+#[cfg(feature = "async")]
+pub use async_api::*;
 
 use crate::{sercom::pad::SomePad, typelevel::Sealed};
 use core::marker::PhantomData;
@@ -502,6 +531,10 @@ pub trait Receive: Capability {}
 /// capability, but not both
 pub trait Simplex: Capability {}
 
+/// Type-level enum representing a UART that is *not* half of a split
+/// [`Duplex`]
+pub trait SingleOwner: Capability {}
+
 /// Marker type representing a UART that has both transmit and receive
 /// capability
 pub enum Duplex {}
@@ -518,6 +551,7 @@ impl Capability for Duplex {
 }
 impl Receive for Duplex {}
 impl Transmit for Duplex {}
+impl SingleOwner for Duplex {}
 
 /// Marker type representing a UART that can only receive
 pub enum Rx {}
@@ -534,6 +568,7 @@ impl Capability for Rx {
 }
 impl Receive for Rx {}
 impl Simplex for Rx {}
+impl SingleOwner for Rx {}
 
 /// Marker type representing a UART that can only transmit
 pub enum Tx {}
@@ -550,6 +585,7 @@ impl Capability for Tx {
 }
 impl Transmit for Tx {}
 impl Simplex for Tx {}
+impl SingleOwner for Tx {}
 
 /// Marker type representing the Rx half of a  [`Duplex`] UART
 pub enum RxDuplex {}
@@ -619,14 +655,14 @@ where
     /// `Capability`
     #[inline]
     fn capability_flags(flags: Flags) -> Flags {
-        flags & unsafe { Flags::from_bits_unchecked(D::FLAG_MASK) }
+        flags & Flags::from_bits_retain(D::FLAG_MASK)
     }
 
     /// Helper method to remove the status flags not pertinent to `Self`'s
     /// `Capability`
     #[inline]
     fn capability_status(status: Status) -> Status {
-        status & unsafe { Status::from_bits_unchecked(D::STATUS_MASK) }
+        status & Status::from_bits_retain(D::STATUS_MASK)
     }
 
     /// Read the interrupt flags
@@ -751,7 +787,7 @@ where
         self.config
             .as_mut()
             .registers
-            .clear_flags(unsafe { Flags::from_bits_unchecked(bit) });
+            .clear_flags(Flags::from_bits_retain(bit));
     }
 
     /// Enable the `CTSIC` interrupt
@@ -761,7 +797,7 @@ where
         self.config
             .as_mut()
             .registers
-            .enable_interrupts(unsafe { Flags::from_bits_unchecked(bit) });
+            .enable_interrupts(Flags::from_bits_retain(bit));
     }
 
     /// Disable the `CTSIC` interrupt
@@ -771,7 +807,7 @@ where
         self.config
             .as_mut()
             .registers
-            .disable_interrupts(unsafe { Flags::from_bits_unchecked(bit) });
+            .disable_interrupts(Flags::from_bits_retain(bit));
     }
 }
 
@@ -910,7 +946,7 @@ where
     /// containing the corresponding [`Flags`] or [`Error`]
     #[inline]
     fn read_flags_errors(&self) -> Result<Flags, Error> {
-        self.read_status().try_into()?;
+        self.read_status().check_bus_error()?;
         Ok(self.read_flags())
     }
 
@@ -926,13 +962,14 @@ where
     #[inline]
     pub fn flush_rx_buffer(&mut self) {
         // TODO Is this a hardware bug???
-        /*
-        usart.ctrlb.modify(|_, w| w.rxen().clear_bit());
-        while usart.syncbusy.read().ctrlb().bit() || usart.ctrlb.read().rxen().bit_is_set() {}
 
-        usart.ctrlb.modify(|_, w| w.rxen().set_bit());
-        while usart.syncbusy.read().ctrlb().bit() || usart.ctrlb.read().rxen().bit_is_clear() {}
-        */
+        // usart.ctrlb.modify(|_, w| w.rxen().clear_bit());
+        // while usart.syncbusy.read().ctrlb().bit() ||
+        // usart.ctrlb.read().rxen().bit_is_set() {}
+
+        // usart.ctrlb.modify(|_, w| w.rxen().set_bit());
+        // while usart.syncbusy.read().ctrlb().bit() ||
+        // usart.ctrlb.read().rxen().bit_is_clear() {}
 
         for _ in 0..=2 {
             let _data = unsafe { self.config.as_mut().registers.read_data() };