+++ boards/feather_m0/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 - Added 'winc' feature for Feather with a WINC1500 Wifi chip
 
+- Add async examples and update examples to match new HAL APIs
+
 # v0.14.0
 
 - Implement `embedded-hal` `1.0` for GPIO, SPI, I2C, UART and fix examples